Uploading MPC example.
The August 22 code has an
issue where it can't reduce the level in the tank
as quickly as in the MATLAB example. This results
in the valve saturating.

The August 22 code also steps from fully closed
to fully open which is not ideal. Such fast valve
actions should try to be avoided.